requirements.txt
Pillow==10.0.0
pandas==2.0.3
numpy==1.25.2
faiss-cpu==1.7.4
streamlit==1.18.1
altair==4.0
openpyxl==3.1.2
scikit-learn==1.5.2